fix: reduced lines in messenger-mocks.ts
